+++ tiger-common/src/test/java/de/gematik/test/tiger/common/config/TestTigerConfiguration.java
@@ -4,16 +4,22 @@
 
 package de.gematik.test.tiger.common.config;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import lombok.SneakyThrows;
+import org.apache.commons.io.FileUtils;
 import org.junit.jupiter.api.Test;
+
+import java.io.File;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 public class TestTigerConfiguration {
 
+    @SneakyThrows
     @Test
-    public void readTestEnvYaml() {
-        assertThat(new TigerConfigurationHelper<TestCfg>().yamlReadOverwriteToConfig(
-            "../tiger-testenv-mgr/src/main/resources/de/gematik/test/tiger/testenvmgr/templates.yaml",
-            "tiger", TestCfg.class))
+    public void readWithTigerConfiguration() {
+        TigerGlobalConfiguration.readFromYaml(
+            FileUtils.readFileToString(new File("../tiger-testenv-mgr/src/main/resources/de/gematik/test/tiger/testenvmgr/templates.yaml")), "tiger");
+        assertThat(TigerGlobalConfiguration.instantiateConfigurationBean(TestCfg.class, "tiger"))
             .extracting(TestCfg::getTemplates)
             .asList()
             .extracting("templateName")