+++ tiger-test-lib/src/main/java/de/gematik/test/tiger/lib/exception/TigerStartupException.java
@@ -8,4 +8,8 @@ public class TigerStartupException extends RuntimeException {
     public TigerStartupException(String message) {
         super(message);
     }
+
+    public TigerStartupException(String message, Exception e) {
+        super(message, e);
+    }
 }