+++ tiger-testenv-mgr/src/main/java/de/gematik/test/tiger/testenvmgr/TigerTestEnvMgr.java
@@ -4,14 +4,12 @@
 
 package de.gematik.test.tiger.testenvmgr;
 
-import static org.awaitility.Awaitility.await;
 import de.gematik.rbellogger.util.RbelAnsiColors;
 import de.gematik.test.tiger.common.Ansi;
-import de.gematik.test.tiger.common.OsEnvironment;
 import de.gematik.test.tiger.common.TokenSubstituteHelper;
 import de.gematik.test.tiger.common.banner.Banner;
+import de.gematik.test.tiger.common.config.TigerGlobalConfiguration;
 import de.gematik.test.tiger.common.config.TigerConfigurationException;
-import de.gematik.test.tiger.common.config.TigerConfigurationHelper;
 import de.gematik.test.tiger.common.config.tigerProxy.TigerProxyConfiguration;
 import de.gematik.test.tiger.common.config.tigerProxy.TigerRoute;
 import de.gematik.test.tiger.common.pki.TigerConfigurationPkiIdentity;
@@ -19,9 +17,16 @@ import de.gematik.test.tiger.proxy.TigerProxy;
 import de.gematik.test.tiger.testenvmgr.config.CfgServer;
 import de.gematik.test.tiger.testenvmgr.config.Configuration;
 import de.gematik.test.tiger.testenvmgr.servers.TigerServer;
+import lombok.Getter;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.IOUtils;
+import org.apache.commons.lang3.StringUtils;
+
 import java.io.*;
 import java.net.InetAddress;
 import java.net.URISyntaxException;
+import java.net.URL;
 import java.net.UnknownHostException;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
@@ -29,10 +34,8 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
-import lombok.Getter;
-import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.io.IOUtils;
-import org.json.JSONObject;
+
+import static org.awaitility.Awaitility.await;
 
 @Slf4j
 @Getter
@@ -40,6 +43,7 @@ public class TigerTestEnvMgr implements ITigerTestEnvMgr {
 
     public static final String HTTP = "http://";
     public static final String HTTPS = "https://";
+    private static final String TIGER_TESTENV_YAML_FILENAME = "tiger-testenv.yaml";
     private final Configuration configuration;
     private final DockerMgr dockerManager;
     private final Map<String, Object> environmentVariables;
@@ -126,17 +130,14 @@ public class TigerTestEnvMgr implements ITigerTestEnvMgr {
         }
     }
 
-    public static void applyTemplates(JSONObject jsonCfg) {
+    private static void readTemplates() {
         // read configuration from file and templates from classpath resource
         try {
-            JSONObject jsonTemplate = TigerConfigurationHelper.yamlStringToJson(
-                IOUtils.toString(Objects.requireNonNull(TigerTestEnvMgr.class.getResource(
-                    "templates.yaml")).toURI(), StandardCharsets.UTF_8));
-            if (jsonCfg.has("servers")) {
-                TigerConfigurationHelper.applyTemplate(
-                    jsonCfg.getJSONObject("servers"), "template",
-                    jsonTemplate.getJSONArray("templates"), "templateName");
-            }
+            final URL templatesUrl = TigerTestEnvMgr.class.getResource("templates.yaml");
+            final String templatesYaml = IOUtils.toString(
+                Objects.requireNonNull(templatesUrl).toURI(),
+                StandardCharsets.UTF_8);
+            TigerGlobalConfiguration.readTemplates(templatesYaml, "tiger", "servers");
         } catch (IOException | URISyntaxException e) {
             throw new TigerConfigurationException("Unable to read templates YAML!", e);
         }
@@ -147,6 +148,40 @@ public class TigerTestEnvMgr implements ITigerTestEnvMgr {
             return InetAddress.getLocalHost().getHostName();
         } catch (UnknownHostException e) {
             return InetAddress.getLoopbackAddress().getHostName();
+        }
+    }
+
+    private static Configuration readConfiguration() {
+        TigerGlobalConfiguration.initialize();
+        readTemplates();
+        readTestenvYaml();
+        final Configuration configuration = TigerGlobalConfiguration.instantiateConfigurationBean(Configuration.class, "tiger");
+        for (CfgServer cfgServer : configuration.getServers().values()) {
+            if (StringUtils.isNotEmpty(cfgServer.getTemplate())) {
+                throw new TigerConfigurationException("Could not resolve template '" + cfgServer.getTemplate() + "'");
+            }
+        }
+        return configuration;
+    }
+
+    private static void readTestenvYaml() {
+        final String configFileLocation = TigerGlobalConfiguration.readString(
+            "TIGER_TESTENV_CFGFILE", "tiger-testenv-" + getComputerName() + ".yaml");
+        var cfgFile = new File(configFileLocation);
+        if (!cfgFile.exists()) {
+            log.warn("Unable to read configuration from {}", cfgFile.getAbsolutePath());
+            cfgFile = new File(TIGER_TESTENV_YAML_FILENAME);
+            if (!cfgFile.exists()) {
+                throw new TigerEnvironmentStartupException("Could not find configuration-file '" + configFileLocation
+                    + "' or '" + TIGER_TESTENV_YAML_FILENAME + "' fallback");
+            }
+        }
+        log.info("Reading configuration from {}...", cfgFile.getAbsolutePath());
+        try {
+            TigerGlobalConfiguration.readFromYaml(FileUtils.readFileToString(cfgFile, StandardCharsets.UTF_8), "tiger");
+        } catch (Exception e) {
+            throw new TigerEnvironmentStartupException(
+                "Error while reading configuration from file '" + cfgFile.getAbsolutePath() + "'", e);
         }
     }
 
@@ -170,31 +205,6 @@ public class TigerTestEnvMgr implements ITigerTestEnvMgr {
             newSet.add(currentPosition);
             cycleChecker(server, newSet);
         }
-    }
-
-    private Configuration readConfiguration() {
-        JSONObject jsonCfg = readConfiguredConfigurationFileAndConvert();
-        applyTemplates(jsonCfg);
-        TigerConfigurationHelper.overwriteWithSysPropsAndEnvVars("TIGER_TESTENV", "tiger.testenv", jsonCfg);
-
-        return new TigerConfigurationHelper<Configuration>()
-            .jsonStringToConfig(jsonCfg.toString(), Configuration.class);
-    }
-
-    private JSONObject readConfiguredConfigurationFileAndConvert() {
-        var cfgFile = new File(OsEnvironment.getAsString(
-            "TIGER_TESTENV_CFGFILE", "tiger-testenv-" + getComputerName() + ".yaml"));
-        if (!cfgFile.exists()) {
-            log.warn("Unable to read configuration from {}", cfgFile.getAbsolutePath());
-            cfgFile = new File("tiger-testenv.yaml");
-            if (!cfgFile.exists()) {
-                throw new TigerEnvironmentStartupException(
-                    "Unable to load configuration from " + System.getProperty("TIGER_TESTENV_CFGFILE") + " or fallback");
-            }
-        }
-        log.info("Reading configuration from {}...", cfgFile.getAbsolutePath());
-        JSONObject jsonCfg = TigerConfigurationHelper.yamlToJson(cfgFile.getAbsolutePath());
-        return jsonCfg;
     }
 
     private void createServerObjects() {